requirements.txt
setuptools
python-decouple
flask
openai
